Update patch set 2
Patch Set 2:

Are you able to try and record the code with lirc on Linux? Maybe that helps to see what the difference between the 2 codes is. It is just an idea though. What TV are you trying to turn on? Do you know the pronto code of it?

Patch-set: 2
Reviewer: Timo Wendt <4098@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Verified=0