Update patch set 3
Patch Set 3: Code-Review+1

I discussed with Arcee, and we should not have settings for double-tap (at least at this time) to change double-tap. This should be merged as is, with the possibility of settings at another time.

Users can always swap Recents action for Menu for long-press home still, although that will not be the default.

Patch-set: 3
Label: Code-Review=+1