Update patch set 3
Patch Set 3:

Look... Google obviously wants you to use Google Now, that doesn't mean we should.
I just don't think "Search" is a more important function than "Menu". The device has 2 keys, and you can make it behave as if it had 3 (which a ton of other devices _do_): Home, Back, and Menu (via double-tap), + Longpress home for recents. A lot of the navigation (and apps) works under the assumption that if keys exist, those are it, so why not default to it? (and don't tell me it's "because that's how the GE behaves")

Patch-set: 3
Reviewer: Ricardo Cerqueira <10@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Verified=0