Update patch set 1
Patch Set 1:

This change has caused the proximity sensor to deactivate during calls on the m7wls or m7spr. Upon placing the phone near the face to hear the call, the screen shuts off and is unable to be turned back on until the other side hangs up the call. The only way to turn it on during the call is by flashing a bright light at the proximity sensor.

Patch-set: 1
Reviewer: Felipe Yang <7223@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Verified=0