Update patch set 1
Patch Set 1: Code-Review-1

This sounds like a problem with the apps or the audio HAL rather than this use case. It makes sense that the speaker use cases attempt to use the back mic more, as the speakers are meant to be in use. I'm going to -1 pending some more investigation.

Patch-set: 1