Update patch set 2
Patch Set 2:

I did some testing and I did not see any negative with this enabled for Phone calls, and actually might make more sense for speakerphone to use the bottom mic rather then the back (usually the phone is placed face up when on a call).  But this change also makes recording of video with camera use the bottom mic and I feel that for video the rear mic would be best... Anyway to use this only for phone and not for camera?

Patch-set: 2
Reviewer: Derek Romano <7259@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Verified=0