Update patch set 2
Patch Set 2:

Adit on a call that is the stock behavior of the GPe 4.3 rom, to use the rear mic, while I agree with you that it makes more sense for calls see my previous comment about camera. I have not tried on a Sense base yet however, wonder what the config is like on sense.

Patch-set: 2