Update patch set 1
Patch Set 1: Code-Review+1

Haven't fully tested(ran the encryption on my device) yet because I don't have the full hour, but his appears to be functional as the option to do it is there now and the button shows up when on the charger.

Looks good to me, but someone else must approve!

Patch-set: 1