Update patch set 1
Patch Set 1: Code-Review+1

This has no negative impact on m7vzw.  Since the default setting in frameworks/base is to set content disposition enabled, and I cannot find a reason this was set to disabled in the first place, I think this should be merged. Any m7 variant which is negatively impacted can re-include this in its own overlay.

Patch-set: 1
Label: Code-Review=+1