Update patch set 1
Patch Set 1:

Christopher: If you want to try the conditional revert you'll need something like this as well as revert all the other recent reverts related to this. Likely also need to revert the updates to the qseecom driver in the kernel if Ethan pulled in the significantly newer CAF source.

Be prepared it might be very broken though.

Patch-set: 1