m7-common: tfa9887: Adding locking for set_mode
It appears multiple threads can be concurrently setting the
amplifier mode.  An easy way to see this is to add pthread_thread_self
into some of the log messages and add a big sleep in the middle of
the initialization code.  You'll see a second initialization begin
to take place.

Change-Id: Ic247fdd8d1c55cd054053b9b7dc38d99a16cb33f