Update patch set 2
Patch Set 2:

I uploaded a new version of the dlx version of this commit.  I, obviously, like it better.  :-)

The "dsp_mode" check is basically an optimization.  If we didn't have it, we would sometimes reconfigure the amp to the same state that it is currently in.

Imagine that we removed the optimization and we are in dsp_mode 0 when two threads start to set it to mode 0 and mode 1.  If the mode 1 thread acquires the KERNEL_LOCK first, then the final result will be that we are in mode 0.

In your version, the mode 0 thread could check the dsp_mode and turn itself into a noop and therefore the final state would be mode 1.

In my version, the dsp_mode check is made under the lock and in this case the final state would be the correct one.

Patch-set: 2