Update patch set 1
Patch Set 1:

I have tested the use case that I would generally use and it seems to work ok.  Since i'm not a big speakerphone user, I can't tell if it's much better.

According to one of the testers though, this patch is causing the other party on the phone to hear themselves speaking.

Patch-set: 1
Label: Code-Review=0