Update patch set 2
Patch Set 2: Code-Review-1

There is a caveat to this. Installation packages will map /system to /dev/block/platform/msm_sdcc.1/by-name/system, but recoveries built without by-name kernel support will not know what to do with that. So, we would be forcing users to update their recoveries as well with this.

-1 until we figure out best route. Maybe we go back to separate fstabs for CM and recovery? Not positive if that will fix the issue, though.

Patch-set: 2