Update patch set 1
Patch Set 1:

The negative is that this is essentially a hack. Android has a method for taking care of this problem: inheritance. It makes sense to aggregate the GSM variants (there is no fundamental difference), but it doesn't make sense to aggregate CDMA with GSM, for reasons to do with the radio stack.

Patch-set: 1