Update patch set 1
Patch Set 1:

I disagree with the idea that it "just happens to work". There's no reason it shouldn't. We're shipping the differences between them. Your GSM variants are already building all the CDMA stuff anyway, just not flipping the switches to enable it (of course hardware prevents it too).

Look at our consolidated Samsungs, they run just fine. The whole idea got a bad wrap to begin with because the qcril merge that happened right when we shipped it. We had been running it for months before that without issue. Sure there were a few hiccups with overlays and apns, but that's since been resolved too.

I'm not saying consolidate M7, that's up to you guys. Its just not as bad of an idea as its being made out to be.

Patch-set: 1