Update patch set 1
Patch Set 1:

I'm with Dan. If the only differences are ril (same ril could support multiple carriers, but don't have the device, so not sure if it does) and properties, then how could this ever not work? It's not luck or random, it's systematically adding any change between variants. This is better for both users (more choices) and developers (easier to build for more devices, less maintenance required).

Dorian, the only thing I would change is naming. The m7 device name is already used, so it won't be obvious if it's a unified build or just international. m7lte could work or m7www, m7unified, m7global, but m7 would be confusing for users, so you would constantly be asked the same questions on XDA.

Patch-set: 1