m7-common: dim the defalut backlight configuration
The default backlight configuration is very bright. Let's turn it down a
notch. This is of course still configurable by the end user so this
shouldn't ruffle too many feathers.

For those who stumble across this commit and are curious how to convert
the percentages in the brightness dialog to the values in this xml file,
this should be easy to interpret in your favorite programming language:

testedPercVals = 0.01*{8, 10, 14, 20, 26, 30, 36, 48, 64, 100};
range = 255;
minLev = 10; <-- defined in this xml file
Round[testedPercVals*(range - minLev) + minLev]
  (result) --> {30, 34, 44, 59, 74, 84, 98, 128, 167, 255}

Here's the new brightness curve: http://imgur.com/O6Tt7zs
(you can pull up the old config on your own device for comparison)

Change-Id: I0601686df4d40fc55094f33e5e2001f9aeecc73b