Update patch set 1
Patch Set 1:

What would cause recovery to zero it?

The purpose is so that we can use the two-stage update process that was added in to 4.4.3 (Updates recovery and then updates the device with the new recovery). We need this to update existing CM Installer devices since the recovery version requirement has changed since the last update.

Patch-set: 1