m7-common: update Voice Call verbs and DMIC EF
Update the default Voice Call verb and a few of the device-specific
Voice Call verbs. Change Speaker DMIC Endfire to a single mic with a new
ACDBID.

Changes come from snd_soc_msm_2x_Fusion3_DMIC_S in stock m7vzw OTA
4.10.605.3.

In testing, the person on the other end of the line said this resolved
issues with their voice echoing back to them. This is not intended to be
a comprehensive update to snd_soc_msm_2x_Fusion3, but rather focuses
only on voice items. There's potentially other changes that could be made
in a separate commit.

Change-Id: Iaa1aeaa27c06e549d059c6042c0d3119f5a6d6f6