Update patch set 2
Patch Set 2:

Space is not a question; there's over 2GB available on /system, so we could even enable dex pre-opt if desired. I'm not as excited as the rest of the maintainers about gello, but if browser isn't receiving further updates (I honestly don't know), then I suppose this is better.

Patch-set: 2